use index array for entities in blockmap (#856)
* use list of entities and link to blockmap using their index. fixes slow
blockmap iterations due to the indirection causing slow CPU loads
because it can't prefetch

* store MonsterMovementSpeed locally

* remove flag function calls

* cast bool to byte since bool is 1 byte and not 4

* use direct monster movement speed

* fix clearing z velocity when against ceiling and z velocity is negative

* simpler handling for falling entities when something underneath moves
fixes bugs with dead cacodemons floating in the air and improves
performance on maps with cacodemon clusters like idumea

* simpler handling for falling entities when something underneath moves
fixes bugs with dead cacodemons being stuck in the air and improves performance on maps like idumea

* improve checks for on entity to remove indirection, use had on entity flag to reduce number of fall checks

+++ Core/World/Bsp/CompactBspTree.cs
@@ -129,7 +129,7 @@ public class CompactBspTree
             ref var node = ref Nodes[nodeIndex];
 
             bool onRight = OnRightNode(x, y, node);
-            int next = *(int*)&onRight;
+            int next = *(byte*)&onRight;
             nodeIndex = node.Children[next];
 
             if ((nodeIndex & BspNodeCompact.IsSubsectorBit) != 0)
